refactor: 코드리뷰 반영 - 함수 분리
